backend-django/pyproject.toml
[tool.poetry]
name = "backend-django"
version = "0.1.0"
description = ""
authors = ["Foshati <amirrezafoshati@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.12.2"
django = "^5.0.6"
psycopg = "^3.2.1"
django-cors-headers = "^4.4.0"
django-filter = "^24.2"
dj-rest-auth = {extras = ["with-social"], version = "^6.0.0"}
python-decouple = "^3.8"
django-ninja = "^1.2.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
